refactor(wow-core): introduce WaitStrategy.FunctionMaterialized interface
- Extract common functionality into a new interface FunctionMaterialized
- Simplify the Materialized class in WaitingForStage
- Improve code organization and reusability
